fix: letting 'jx get vault-config' handle external Vault instances as well
- moving vault get commands into dedicated packages
- adjusting 'jx get vault' to also display system vault
